+++ aeon/distances/elastic/_alignment_paths.py
