+++ aeon/distances/elastic/_bounding_matrix.py
