+++ aeon/distances/elastic/_ddtw.py
@@ -8,9 +8,13 @@ import numpy as np
 from numba import njit
 from numba.typed import List as NumbaList
 
-from aeon.distances._alignment_paths import compute_min_return_path
-from aeon.distances._dtw import _dtw_cost_matrix, _dtw_distance, create_bounding_matrix
 from aeon.distances._utils import _convert_to_list, _is_multivariate
+from aeon.distances.elastic._alignment_paths import compute_min_return_path
+from aeon.distances.elastic._dtw import (
+    _dtw_cost_matrix,
+    _dtw_distance,
+    create_bounding_matrix,
+)
 
 
 @njit(cache=True, fastmath=True)
@@ -399,7 +403,7 @@ def average_of_slope(q: np.ndarray) -> np.ndarray:
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances._ddtw import average_of_slope
+    >>> from aeon.distances.elastic._ddtw import average_of_slope
     >>> q = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
     >>> average_of_slope(q)
     array([[1., 1., 1., 1., 1., 1., 1., 1.]])
