+++ aeon/distances/elastic/_dtw.py
@@ -1,4 +1,4 @@
-r"""Amercing dynamic time warping (ADTW) between two time series."""
+r"""Dynamic time warping (DTW) between two time series."""
 
 __maintainer__ = []
 
@@ -8,27 +8,60 @@ import numpy as np
 from numba import njit
 from numba.typed import List as NumbaList
 
-from aeon.distances._alignment_paths import compute_min_return_path
-from aeon.distances._bounding_matrix import create_bounding_matrix
 from aeon.distances._squared import _univariate_squared_distance
 from aeon.distances._utils import _convert_to_list, _is_multivariate
+from aeon.distances.elastic._alignment_paths import compute_min_return_path
+from aeon.distances.elastic._bounding_matrix import create_bounding_matrix
 
 
 @njit(cache=True, fastmath=True)
-def adtw_distance(
+def dtw_distance(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
-    warp_penalty: float = 1.0,
 ) -> float:
-    r"""Compute the ADTW distance between two time series.
+    r"""Compute the DTW distance between two time series.
 
-    Amercing Dynamic Time Warping (ADTW) [1]_ is a variant of DTW that uses a
-    explicit warping penalty to encourage or discourage warping. The warping
-    penalty is a constant value that is added to the cost of warping. A high
-    value will encourage the algorithm to warp less and if the value is low warping
-    is more likely.
+    DTW is the most widely researched and used elastic distance measure. It mitigates
+    distortions in the time axis by realligning (warping) the series to best match
+    each other. A good background into DTW can be found in [1]_. For two series,
+    possibly of unequal length,
+    :math:`\mathbf{x}=\{x_1,x_2,\ldots,x_n\}` and
+    :math:`\mathbf{y}=\{y_1,y_2, \ldots,y_m\}` DTW first calculates
+    :math:`M(\mathbf{x},\mathbf{y})`, the :math:`n \times m`
+    pointwise distance matrix between series :math:`\mathbf{x}` and :math:`\mathbf{y}`,
+    where :math:`M_{i,j}=   (x_i-y_j)^2`.
+
+    A warping path
+
+    .. math::
+        P = <(e_1, f_1), (e_2, f_2), \ldots, (e_s, f_s)>
+
+    is a set of pairs of indices that  define a traversal of matrix :math:`M`. A
+    valid warping path must start at location :math:`(1,1)` and end at point :math:`(
+    n,m)` and not backtrack, i.e. :math:`0 \leq e_{i+1}-e_{i} \leq 1` and :math:`0
+    \leq f_{i+1}- f_i \leq 1` for all :math:`1< i < m`.
+
+    The DTW distance between series is the path through :math:`M` that minimizes the
+    total distance. The distance for any path :math:`P` of length :math:`s` is
+
+    .. math::
+        D_P(\mathbf{x},\mathbf{y}, M) =\sum_{i=1}^s M_{e_i,f_i}
+
+    If :math:`\mathcal{P}` is the space of all possible paths, the DTW path :math:`P^*`
+    is the path that has the minimum distance, hence the DTW distance between series is
+
+    .. math::
+        d_{dtw}(\mathbf{x}, \mathbf{x}) =D_{P*}(\mathbf{x},\mathbf{x}, M).
+
+    The optimal warping path :math:`P^*` can be found exactly through a dynamic
+    programming formulation. This can be a time consuming operation, and it is common to
+    put a restriction on the amount of warping allowed. This is implemented through
+    the bounding_matrix structure, that supplies a mask for allowable warpings.
+    The most common bounding strategies include the Sakoe-Chiba band [2]_. The width
+    of the allowed warping is controlled through the ``window`` parameter
+    which sets the maximum proportion of warping allowed.
 
     Parameters
     ----------
@@ -42,16 +75,15 @@ def adtw_distance(
         The window to use for the bounding matrix. If None, no bounding matrix
         is used. window is a percentage deviation, so if ``window = 0.1`` then
         10% of the series length is the max warping allowed.
+        is used.
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
-        Itakura parallelogram on the bounding matrix. Must be between 0.0 and 1.0
-    warp_penalty: float, default=1.0
-        Penalty for warping. A high value will mean less warping.
+        Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
 
     Returns
     -------
     float
-        ADTW distance between x and y, minimum value 0.
+        DTW distance between x and y, minimum value 0.
 
     Raises
     ------
@@ -60,21 +92,25 @@ def adtw_distance(
 
     References
     ----------
-    .. [1] Matthieu Herrmann, Geoffrey I. Webb: Amercing: An intuitive and effective
-    constraint for dynamic time warping, Pattern Recognition, Volume 137, 2023.
+    .. [1] Ratanamahatana C and Keogh E.: Three myths about dynamic time warping data
+    mining, Proceedings of 5th SIAM International Conference on Data Mining, 2005.
+
+    .. [2] Sakoe H. and Chiba S.: Dynamic programming algorithm optimization for
+    spoken word recognition. IEEE Transactions on Acoustics, Speech, and Signal
+    Processing 26(1):43–49, 1978.
 
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import adtw_distance
+    >>> from aeon.distances import dtw_distance
     >>> x = np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
     >>> y = np.array([11, 12, 13, 14, 15, 16, 17, 18, 19, 20])
-    >>> adtw_distance(x, y) # 1D series
-    783.0
+    >>> dtw_distance(x, y) # 1D series
+    768.0
     >>> x = np.array([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10], [0, 1, 0, 2, 0]])
     >>> y = np.array([[11, 12, 13, 14],[7, 8, 9, 20],[1, 3, 4, 5]] )
-    >>> adtw_distance(x, y) # 2D series with 3 channels, unequal length
-    565.0
+    >>> dtw_distance(x, y) # 2D series with 3 channels, unequal length
+    564.0
     """
     if x.ndim == 1 and y.ndim == 1:
         _x = x.reshape((1, x.shape[0]))
@@ -82,24 +118,26 @@ def adtw_distance(
         bounding_matrix = create_bounding_matrix(
             _x.shape[1], _y.shape[1], window, itakura_max_slope
         )
-        return _adtw_distance(_x, _y, bounding_matrix, warp_penalty)
+        return _dtw_distance(_x, _y, bounding_matrix)
     if x.ndim == 2 and y.ndim == 2:
         bounding_matrix = create_bounding_matrix(
             x.shape[1], y.shape[1], window, itakura_max_slope
         )
-        return _adtw_distance(x, y, bounding_matrix, warp_penalty)
+        return _dtw_distance(x, y, bounding_matrix)
     raise ValueError("x and y must be 1D or 2D")
 
 
 @njit(cache=True, fastmath=True)
-def adtw_cost_matrix(
+def dtw_cost_matrix(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
-    warp_penalty: float = 1.0,
 ) -> np.ndarray:
-    r"""Compute the ADTW cost matrix between two time series.
+    r"""Compute the DTW cost matrix between two time series.
+
+    The cost matrix is the pairwise Euclidean distance between all points
+    :math:`M_{i,j}=(x_i-x_j)^2`. It is used in the DTW path calculations.
 
     Parameters
     ----------
@@ -117,13 +155,11 @@ def adtw_cost_matrix(
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
-    warp_penalty: float, default=1.0
-        Penalty for warping. A high value will mean less warping.
 
     Returns
     -------
     np.ndarray (n_timepoints, m_timepoints)
-        adtw cost matrix between x and y.
+        dtw cost matrix between x and y.
 
     Raises
     ------
@@ -133,21 +169,20 @@ def adtw_cost_matrix(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import adtw_cost_matrix
+    >>> from aeon.distances import dtw_cost_matrix
     >>> x = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
     >>> y = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
-    >>> adtw_cost_matrix(x, y)
-    array([[  0.,   2.,   7.,  17.,  34.,  60.,  97., 147., 212., 294.],
-           [  2.,   0.,   2.,   7.,  17.,  34.,  60.,  97., 147., 212.],
-           [  7.,   2.,   0.,   2.,   7.,  17.,  34.,  60.,  97., 147.],
-           [ 17.,   7.,   2.,   0.,   2.,   7.,  17.,  34.,  60.,  97.],
-           [ 34.,  17.,   7.,   2.,   0.,   2.,   7.,  17.,  34.,  60.],
-           [ 60.,  34.,  17.,   7.,   2.,   0.,   2.,   7.,  17.,  34.],
-           [ 97.,  60.,  34.,  17.,   7.,   2.,   0.,   2.,   7.,  17.],
-           [147.,  97.,  60.,  34.,  17.,   7.,   2.,   0.,   2.,   7.],
-           [212., 147.,  97.,  60.,  34.,  17.,   7.,   2.,   0.,   2.],
-           [294., 212., 147.,  97.,  60.,  34.,  17.,   7.,   2.,   0.]])
-
+    >>> dtw_cost_matrix(x, y)
+    array([[  0.,   1.,   5.,  14.,  30.,  55.,  91., 140., 204., 285.],
+           [  1.,   0.,   1.,   5.,  14.,  30.,  55.,  91., 140., 204.],
+           [  5.,   1.,   0.,   1.,   5.,  14.,  30.,  55.,  91., 140.],
+           [ 14.,   5.,   1.,   0.,   1.,   5.,  14.,  30.,  55.,  91.],
+           [ 30.,  14.,   5.,   1.,   0.,   1.,   5.,  14.,  30.,  55.],
+           [ 55.,  30.,  14.,   5.,   1.,   0.,   1.,   5.,  14.,  30.],
+           [ 91.,  55.,  30.,  14.,   5.,   1.,   0.,   1.,   5.,  14.],
+           [140.,  91.,  55.,  30.,  14.,   5.,   1.,   0.,   1.,   5.],
+           [204., 140.,  91.,  55.,  30.,  14.,   5.,   1.,   0.,   1.],
+           [285., 204., 140.,  91.,  55.,  30.,  14.,   5.,   1.,   0.]])
     """
     if x.ndim == 1 and y.ndim == 1:
         _x = x.reshape((1, x.shape[0]))
@@ -155,27 +190,23 @@ def adtw_cost_matrix(
         bounding_matrix = create_bounding_matrix(
             _x.shape[1], _y.shape[1], window, itakura_max_slope
         )
-        return _adtw_cost_matrix(_x, _y, bounding_matrix, warp_penalty)
+        return _dtw_cost_matrix(_x, _y, bounding_matrix)
     if x.ndim == 2 and y.ndim == 2:
         bounding_matrix = create_bounding_matrix(
             x.shape[1], y.shape[1], window, itakura_max_slope
         )
-        return _adtw_cost_matrix(x, y, bounding_matrix, warp_penalty)
+        return _dtw_cost_matrix(x, y, bounding_matrix)
     raise ValueError("x and y must be 1D or 2D")
 
 
 @njit(cache=True, fastmath=True)
-def _adtw_distance(
-    x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray, warp_penalty: float
-) -> float:
-    return _adtw_cost_matrix(x, y, bounding_matrix, warp_penalty)[
-        x.shape[1] - 1, y.shape[1] - 1
-    ]
+def _dtw_distance(x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray) -> float:
+    return _dtw_cost_matrix(x, y, bounding_matrix)[x.shape[1] - 1, y.shape[1] - 1]
 
 
 @njit(cache=True, fastmath=True)
-def _adtw_cost_matrix(
-    x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray, warp_penalty: float
+def _dtw_cost_matrix(
+    x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray
 ) -> np.ndarray:
     x_size = x.shape[1]
     y_size = y.shape[1]
@@ -188,22 +219,37 @@ def _adtw_cost_matrix(
                 cost_matrix[i + 1, j + 1] = _univariate_squared_distance(
                     x[:, i], y[:, j]
                 ) + min(
-                    cost_matrix[i, j + 1] + warp_penalty,
-                    cost_matrix[i + 1, j] + warp_penalty,
+                    cost_matrix[i, j + 1],
+                    cost_matrix[i + 1, j],
                     cost_matrix[i, j],
                 )
 
     return cost_matrix[1:, 1:]
 
 
-def adtw_pairwise_distance(
+def dtw_pairwise_distance(
     X: Union[np.ndarray, list[np.ndarray]],
     y: Optional[Union[np.ndarray, list[np.ndarray]]] = None,
     window: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
-    warp_penalty: float = 1.0,
 ) -> np.ndarray:
-    r"""Compute the ADTW pairwise distance between a set of time series.
+    r"""Compute the DTW pairwise distance between a set of time series.
+
+    By default, this takes a collection of :math:`n` time series :math:`X` and returns a
+    matrix
+    :math:`D` where :math:`D_{i,j}` is the DTW distance between the :math:`i^{th}`
+    and the :math:`j^{th}` series in :math:`X`. If :math:`X` is 2 dimensional,
+    it is assumed to be a collection of univariate series with shape ``(n_cases,
+    n_timepoints)``. If it is 3 dimensional, it is assumed to be shape ``(n_cases,
+    n_channels, n_timepoints)``.
+
+    This function has an optional argument, :math:`y`, to allow calculation of the
+    distance matrix between :math:`X` and one or more series stored in :math:`y`. If
+    :math:`y` is 1 dimensional, we assume it is a single univariate series and the
+    distance matrix returned is shape ``(n_cases,1)``. If it is 2D, we assume it
+    is a collection of univariate series with shape ``(m_cases, m_timepoints)``
+    and the distance ``(n_cases,m_cases)``. If it is 3 dimensional,
+    it is assumed to be shape ``(m_cases, m_channels, m_timepoints)``.
 
     Parameters
     ----------
@@ -213,7 +259,7 @@ def adtw_pairwise_distance(
     y : np.ndarray or List of np.ndarray or None, default=None
         A single series or a collection of time series of shape ``(m_timepoints,)`` or
         ``(m_cases, m_timepoints)`` or ``(m_cases, m_channels, m_timepoints)``.
-        If None, then the adtw pairwise distance between the instances of X is
+        If None, then the dtw pairwise distance between the instances of X is
         calculated.
     window : float or None, default=None
         The window to use for the bounding matrix. If None, no bounding matrix
@@ -221,15 +267,11 @@ def adtw_pairwise_distance(
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
-    warp_penalty: float, default=1.0
-        Penalty for warping. A high value will mean less warping.
-        warp less and if value is low then will encourage algorithm to warp
-        more.
 
     Returns
     -------
     np.ndarray
-        ADTW pairwise matrix between the instances of X of shape
+        DTW pairwise matrix between the instances of X of shape
         ``(n_cases, n_cases)`` or between X and y of shape ``(n_cases,
         n_cases)``.
 
@@ -241,56 +283,53 @@ def adtw_pairwise_distance(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import adtw_pairwise_distance
+    >>> from aeon.distances import dtw_pairwise_distance
     >>> # Distance between each time series in a collection of time series
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
-    >>> adtw_pairwise_distance(X)
-    array([[  0.,  27., 108.],
-           [ 27.,   0.,  27.],
-           [108.,  27.,   0.]])
+    >>> dtw_pairwise_distance(X)
+    array([[  0.,  26., 108.],
+           [ 26.,   0.,  26.],
+           [108.,  26.,   0.]])
 
     >>> # Distance between two collections of time series
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
     >>> y = np.array([[[11, 12, 13]],[[14, 15, 16]], [[17, 18, 19]]])
-    >>> adtw_pairwise_distance(X, y)
+    >>> dtw_pairwise_distance(X, y)
     array([[300., 507., 768.],
            [147., 300., 507.],
            [ 48., 147., 300.]])
 
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
     >>> y_univariate = np.array([11, 12, 13])
-    >>> adtw_pairwise_distance(X, y_univariate)
+    >>> dtw_pairwise_distance(X, y_univariate)
     array([[300.],
            [147.],
            [ 48.]])
 
     >>> # Distance between each TS in a collection of unequal-length time series
     >>> X = [np.array([1, 2, 3]), np.array([4, 5, 6, 7]), np.array([8, 9, 10, 11, 12])]
-    >>> adtw_pairwise_distance(X)
-    array([[  0.,  44., 294.],
-           [ 44.,   0.,  87.],
-           [294.,  87.,   0.]])
+    >>> dtw_pairwise_distance(X)
+    array([[  0.,  42., 292.],
+           [ 42.,   0.,  83.],
+           [292.,  83.,   0.]])
     """
     multivariate_conversion = _is_multivariate(X, y)
     _X, unequal_length = _convert_to_list(X, "X", multivariate_conversion)
+
     if y is None:
         # To self
-        return _adtw_pairwise_distance(
-            _X, window, itakura_max_slope, warp_penalty, unequal_length
-        )
-
+        return _dtw_pairwise_distance(_X, window, itakura_max_slope, unequal_length)
     _y, unequal_length = _convert_to_list(y, "y", multivariate_conversion)
-    return _adtw_from_multiple_to_multiple_distance(
-        _X, _y, window, itakura_max_slope, warp_penalty, unequal_length
+    return _dtw_from_multiple_to_multiple_distance(
+        _X, _y, window, itakura_max_slope, unequal_length
     )
 
 
 @njit(cache=True, fastmath=True)
-def _adtw_pairwise_distance(
+def _dtw_pairwise_distance(
     X: NumbaList[np.ndarray],
     window: Optional[float],
     itakura_max_slope: Optional[float],
-    warp_penalty: float,
     unequal_length: bool,
 ) -> np.ndarray:
     n_cases = len(X)
@@ -308,19 +347,18 @@ def _adtw_pairwise_distance(
                 bounding_matrix = create_bounding_matrix(
                     x1.shape[1], x2.shape[1], window, itakura_max_slope
                 )
-            distances[i, j] = _adtw_distance(x1, x2, bounding_matrix, warp_penalty)
+            distances[i, j] = _dtw_distance(x1, x2, bounding_matrix)
             distances[j, i] = distances[i, j]
 
     return distances
 
 
 @njit(cache=True, fastmath=True)
-def _adtw_from_multiple_to_multiple_distance(
+def _dtw_from_multiple_to_multiple_distance(
     x: NumbaList[np.ndarray],
     y: NumbaList[np.ndarray],
     window: Optional[float],
     itakura_max_slope: Optional[float],
-    warp_penalty: float,
     unequal_length: bool,
 ) -> np.ndarray:
     n_cases = len(x)
@@ -338,19 +376,18 @@ def _adtw_from_multiple_to_multiple_distance(
                 bounding_matrix = create_bounding_matrix(
                     x1.shape[1], y1.shape[1], window, itakura_max_slope
                 )
-            distances[i, j] = _adtw_distance(x1, y1, bounding_matrix, warp_penalty)
+            distances[i, j] = _dtw_distance(x1, y1, bounding_matrix)
     return distances
 
 
 @njit(cache=True, fastmath=True)
-def adtw_alignment_path(
+def dtw_alignment_path(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
-    warp_penalty: float = 1.0,
 ) -> tuple[list[tuple[int, int]], float]:
-    """Compute the ADTW alignment path between two time series.
+    """Compute the DTW alignment path between two time series.
 
     Parameters
     ----------
@@ -364,8 +401,6 @@ def adtw_alignment_path(
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
-    warp_penalty: float, default=1.0
-        Penalty for warping. A high value will mean less warping.
 
     Returns
     -------
@@ -374,7 +409,7 @@ def adtw_alignment_path(
         of the index in x and the index in y that have the best alignment according
         to the cost matrix.
     float
-        The ADTW distance betweeen the two time series.
+        The DTW distance betweeen the two time series.
 
     Raises
     ------
@@ -384,13 +419,13 @@ def adtw_alignment_path(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import adtw_alignment_path
+    >>> from aeon.distances import dtw_alignment_path
     >>> x = np.array([[1, 2, 3, 6]])
     >>> y = np.array([[1, 2, 3, 4]])
-    >>> adtw_alignment_path(x, y)
+    >>> dtw_alignment_path(x, y)
     ([(0, 0), (1, 1), (2, 2), (3, 3)], 4.0)
     """
-    cost_matrix = adtw_cost_matrix(x, y, window, itakura_max_slope, warp_penalty)
+    cost_matrix = dtw_cost_matrix(x, y, window, itakura_max_slope)
     return (
         compute_min_return_path(cost_matrix),
         cost_matrix[x.shape[-1] - 1, y.shape[-1] - 1],
