+++ aeon/distances/elastic/_edr.py
@@ -1,4 +1,4 @@
-r"""Longest common subsequence (LCSS) between two time series."""
+"""Edit distance for real sequences (EDR) between two time series."""
 
 __maintainer__ = []
 
@@ -8,57 +8,46 @@ import numpy as np
 from numba import njit
 from numba.typed import List as NumbaList
 
-from aeon.distances._alignment_paths import compute_lcss_return_path
-from aeon.distances._bounding_matrix import create_bounding_matrix
 from aeon.distances._euclidean import _univariate_euclidean_distance
 from aeon.distances._utils import _convert_to_list, _is_multivariate
+from aeon.distances.elastic._alignment_paths import compute_min_return_path
+from aeon.distances.elastic._bounding_matrix import create_bounding_matrix
 
 
 @njit(cache=True, fastmath=True)
-def lcss_distance(
+def edr_distance(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
-    epsilon: float = 1.0,
+    epsilon: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
 ) -> float:
-    r"""Return the LCSS distance between x and y.
+    r"""Compute the EDR distance between two time series.
 
-    The LCSS distance for time series is based on the solution to the
-    longest common subsequence problem in pattern matching [1]_. The typical problem
-    is to find the longest subsequence that is common to two discrete series based on
-    the edit distance. This approach can be extended to consider real-valued time series
-    by using a distance threshold epsilon, which defines the maximum difference
-    between a pair of values that is allowed for them to be considered a match.
-    LCSS finds the optimal alignment between two series by find the greatest number
-    of matching pairs. The LCSS distance uses a matrix :math:`L` that records the
-    sequence of matches over valid warpings. For two series :math:`a = a_1,... a_n`
-    and :math:`b = b_1,... b_m, L'` is found by iterating over all valid windows (i.e.
-    where bounding_matrix is not infinity, which by default is the constant band
-    :math:`|i-j|<w*m`, where :math:`w` is the window parameter value and :math:`m` is
-    series length), then calculating
+    Edit Distance on Real Sequences (EDR) was proposed as an adaptation of standard
+    edit distance on discrete sequences in [1]_, specifically for distances between
+    trajectories. Like LCSS, EDR uses a distance threshold to define when two
+    elements of a series match. However, rather than simply count matches and look
+    for the longest sequence, EDR applies a (constant) penalty for non-matching elements
+    where gaps are inserted to create an optimal alignment.
 
-    :: math..
-        if(|a_i - b_j| < \espilon) \\
-            & L_{i,j} = L_{i-1,j-1}+1 \\
+    .. math::
+        if \;\; |ai − bj | < ϵ\\
+            c &= 0\\
         else\\
-            &L_{i,j} = \max(L_{i,j-1}, L_{i-1,j})\\
+            c &= 1\\
+        match  &=  D_{i-1,j-1}+ c)\\
+        delete &=   D_{i-1,j}+ d({x_{i},g})\\
+        insert &=  D_{i-1,j-1}+ d({g,y_{j}})\\
+        D_{i,j} &= min(match,insert, delete)
 
-    The distance is an inverse function of the longest common subsequence
-    length, :math:`L_{n,m}`.
+    EDR computes the minimum number of elements (as a percentage) that must be removed
+    from x and y so that the sum of the distance between the remaining signal elements
+    lies within the tolerance (epsilon). EDR was originally proposed in [1]_.
 
-    :: math..
-        d_{LCSS}({\bfx,by}) = 1- \frac{L_{n,m}.
-
-    Note that series a and b need not be equal length.
-
-    LCSS attempts to find the longest common sequence between two time series and
-    returns a value that is the percentage that longest common sequence assumes.
-    Originally present in [1]_, LCSS is computed by matching indexes that are
-    similar up until a defined threshold (epsilon).
-
-    The value returned will be between 0.0 and 1.0, where 0.0 means the two time series
-    are exactly the same and 1.0 means they are complete opposites.
+    The value returned will be between 0 and 1 per time series. The value will
+    represent as a percentage of elements that must be removed for the time series to
+    be an exact match.
 
     Parameters
     ----------
@@ -71,9 +60,10 @@ def lcss_distance(
     window : float, default=None
         The window to use for the bounding matrix. If None, no bounding matrix
         is used.
-    epsilon : float, default=1.
+    epsilon : float, default=None
         Matching threshold to determine if two subsequences are considered close
-        enough to be considered 'common'. The default is 1.
+        enough to be considered 'common'. If not specified as per the original paper
+        epsilon is set to a quarter of the maximum standard deviation.
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
@@ -81,26 +71,28 @@ def lcss_distance(
     Returns
     -------
     float
-        The LCSS distance between x and y.
+        EDR distance between x and y.
 
     Raises
     ------
     ValueError
         If x and y are not 1D or 2D arrays.
 
+
     References
     ----------
-    .. [1] M. Vlachos, D. Gunopoulos, and G. Kollios. 2002. "Discovering
-        Similar Multidimensional Trajectories", In Proceedings of the
-        18th International Conference on Data Engineering (ICDE '02).
+    .. [1] Chen L, Ozsu MT, Oria V: Robust and fast similarity search for moving
+    object trajectories. In: Proceedings of the ACM SIGMOD International Conference
+    on Management of Data, 2005
 
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import lcss_distance
+    >>> from aeon.distances import edr_distance
     >>> x = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
     >>> y = np.array([[11, 12, 13, 14, 15, 16, 17, 18, 19, 20]])
-    >>> dist = lcss_distance(x, y)
+    >>> edr_distance(x, y)
+    1.0
     """
     if x.ndim == 1 and y.ndim == 1:
         _x = x.reshape((1, x.shape[0]))
@@ -108,24 +100,24 @@ def lcss_distance(
         bounding_matrix = create_bounding_matrix(
             _x.shape[1], _y.shape[1], window, itakura_max_slope
         )
-        return _lcss_distance(_x, _y, bounding_matrix, epsilon)
+        return _edr_distance(_x, _y, bounding_matrix, epsilon)
     if x.ndim == 2 and y.ndim == 2:
         bounding_matrix = create_bounding_matrix(
             x.shape[1], y.shape[1], window, itakura_max_slope
         )
-        return _lcss_distance(x, y, bounding_matrix, epsilon)
+        return _edr_distance(x, y, bounding_matrix, epsilon)
     raise ValueError("x and y must be 1D or 2D")
 
 
 @njit(cache=True, fastmath=True)
-def lcss_cost_matrix(
+def edr_cost_matrix(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
-    epsilon: float = 1.0,
+    epsilon: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
 ) -> np.ndarray:
-    r"""Return the LCSS cost matrix between x and y.
+    """Compute the EDR cost matrix between two time series.
 
     Parameters
     ----------
@@ -135,20 +127,18 @@ def lcss_cost_matrix(
     y : np.ndarray
         Second time series, either univariate, shape ``(n_timepoints,)``, or
         multivariate, shape ``(n_channels, n_timepoints)``.
-    window : float, default=None
-        The window to use for the bounding matrix. If None, no bounding matrix
-        is used.
-    epsilon : float, default=1.
+    epsilon : float, default=None
         Matching threshold to determine if two subsequences are considered close
-        enough to be considered 'common'. The default is 1.
+        enough to be considered 'common'. If not specified as per the original paper
+        epsilon is set to a quarter of the maximum standard deviation.
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
 
     Returns
     -------
-    np.ndarray
-        The LCSS cost matrix between x and y.
+    np.ndarray (n_timepoints, m_timepoints)
+        EDR cost matrix between x and y.
 
     Raises
     ------
@@ -158,21 +148,20 @@ def lcss_cost_matrix(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import lcss_cost_matrix
+    >>> from aeon.distances import edr_cost_matrix
     >>> x = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
     >>> y = np.array([[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]])
-    >>> lcss_cost_matrix(x, y)
-    array([[ 0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
-           [ 0.,  1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.,  1.],
-           [ 0.,  1.,  2.,  2.,  2.,  2.,  2.,  2.,  2.,  2.,  2.],
-           [ 0.,  1.,  2.,  3.,  3.,  3.,  3.,  3.,  3.,  3.,  3.],
-           [ 0.,  1.,  2.,  3.,  4.,  4.,  4.,  4.,  4.,  4.,  4.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  5.,  5.,  5.,  5.,  5.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  6.,  6.,  6.,  6.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  7.,  7.,  7.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  8.,  8.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9.,  9.],
-           [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10.]])
+    >>> edr_cost_matrix(x, y)
+    array([[0., 1., 1., 1., 1., 1., 1., 1., 1., 1.],
+           [1., 0., 1., 2., 2., 2., 2., 2., 2., 2.],
+           [1., 1., 0., 1., 2., 3., 3., 3., 3., 3.],
+           [1., 2., 1., 0., 1., 2., 3., 4., 4., 4.],
+           [1., 2., 2., 1., 0., 1., 2., 3., 4., 5.],
+           [1., 2., 3., 2., 1., 0., 1., 2., 3., 4.],
+           [1., 2., 3., 3., 2., 1., 0., 1., 2., 3.],
+           [1., 2., 3., 4., 3., 2., 1., 0., 1., 2.],
+           [1., 2., 3., 4., 4., 3., 2., 1., 0., 1.],
+           [1., 2., 3., 4., 5., 4., 3., 2., 1., 0.]])
     """
     if x.ndim == 1 and y.ndim == 1:
         _x = x.reshape((1, x.shape[0]))
@@ -180,55 +169,73 @@ def lcss_cost_matrix(
         bounding_matrix = create_bounding_matrix(
             _x.shape[1], _y.shape[1], window, itakura_max_slope
         )
-        return _lcss_cost_matrix(_x, _y, bounding_matrix, epsilon)
+        return _edr_cost_matrix(_x, _y, bounding_matrix, epsilon)
     if x.ndim == 2 and y.ndim == 2:
         bounding_matrix = create_bounding_matrix(
             x.shape[1], y.shape[1], window, itakura_max_slope
         )
-        return _lcss_cost_matrix(x, y, bounding_matrix, epsilon)
+        return _edr_cost_matrix(x, y, bounding_matrix, epsilon)
     raise ValueError("x and y must be 1D or 2D")
 
 
 @njit(cache=True, fastmath=True)
-def _lcss_distance(
-    x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray, epsilon: float
+def _edr_distance(
+    x: np.ndarray,
+    y: np.ndarray,
+    bounding_matrix: np.ndarray,
+    epsilon: Optional[float] = None,
 ) -> float:
-    distance = _lcss_cost_matrix(x, y, bounding_matrix, epsilon)[x.shape[1], y.shape[1]]
-    distance = 1 - (float(distance / min(x.shape[1], y.shape[1])))
-    if distance < 0.0:
-        return 0.0
-    return distance
+    distance = _edr_cost_matrix(x, y, bounding_matrix, epsilon)[
+        x.shape[1] - 1, y.shape[1] - 1
+    ]
+    return float(distance / max(x.shape[1], y.shape[1]))
 
 
 @njit(cache=True, fastmath=True)
-def _lcss_cost_matrix(
-    x: np.ndarray, y: np.ndarray, bounding_matrix: np.ndarray, epsilon
+def _edr_cost_matrix(
+    x: np.ndarray,
+    y: np.ndarray,
+    bounding_matrix: np.ndarray,
+    epsilon: Optional[float] = None,
 ) -> np.ndarray:
     x_size = x.shape[1]
     y_size = y.shape[1]
+    if epsilon is None:
+        epsilon = float(max(np.std(x), np.std(y))) / 4
 
-    cost_matrix = np.zeros((x_size + 1, y_size + 1))
+    cost_matrix = np.full((x_size + 1, y_size + 1), np.inf)
+
+    for i in range(1, x_size + 1):
+        if bounding_matrix[i - 1, 0]:
+            cost_matrix[i, 0] = 0
+    for j in range(y_size):
+        if bounding_matrix[0, j - 1]:
+            cost_matrix[0, j] = 0
+    cost_matrix[0, 0] = 0
 
     for i in range(1, x_size + 1):
         for j in range(1, y_size + 1):
             if bounding_matrix[i - 1, j - 1]:
-                if _univariate_euclidean_distance(x[:, i - 1], y[:, j - 1]) <= epsilon:
-                    cost_matrix[i, j] = 1 + cost_matrix[i - 1, j - 1]
+                if _univariate_euclidean_distance(x[:, i - 1], y[:, j - 1]) < epsilon:
+                    cost = 0
                 else:
-                    cost_matrix[i, j] = max(
-                        cost_matrix[i, j - 1], cost_matrix[i - 1, j]
-                    )
-    return cost_matrix
+                    cost = 1
+                cost_matrix[i, j] = min(
+                    cost_matrix[i - 1, j - 1] + cost,
+                    cost_matrix[i - 1, j] + 1,
+                    cost_matrix[i, j - 1] + 1,
+                )
+    return cost_matrix[1:, 1:]
 
 
-def lcss_pairwise_distance(
+def edr_pairwise_distance(
     X: Union[np.ndarray, list[np.ndarray]],
     y: Optional[Union[np.ndarray, list[np.ndarray]]] = None,
     window: Optional[float] = None,
-    epsilon: float = 1.0,
+    epsilon: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
 ) -> np.ndarray:
-    """Compute the LCSS pairwise distance between a set of time series.
+    """Compute the pairwise EDR distance between a set of time series.
 
     Parameters
     ----------
@@ -238,14 +245,15 @@ def lcss_pairwise_distance(
     y : np.ndarray or List of np.ndarray or None, default=None
         A single series or a collection of time series of shape ``(m_timepoints,)`` or
         ``(m_cases, m_timepoints)`` or ``(m_cases, m_channels, m_timepoints)``.
-        If None, then the lcss pairwise distance between the instances of X is
+        If None, then the edr pairwise distance between the instances of X is
         calculated.
     window : float, default=None
         The window to use for the bounding matrix. If None, no bounding matrix
         is used.
-    epsilon : float, default=1.
+    epsilon : float, default=None
         Matching threshold to determine if two subsequences are considered close
-        enough to be considered 'common'. The default is 1.
+        enough to be considered 'common'. If not specified as per the original paper
+        epsilon is set to a quarter of the maximum standard deviation.
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
@@ -253,7 +261,7 @@ def lcss_pairwise_distance(
     Returns
     -------
     np.ndarray (n_cases, n_cases)
-        LCSS pairwise matrix between the instances of X.
+        EDR pairwise matrix between the instances of X.
 
     Raises
     ------
@@ -264,59 +272,62 @@ def lcss_pairwise_distance(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import lcss_pairwise_distance
+    >>> from aeon.distances import edr_pairwise_distance
     >>> # Distance between each time series in a collection of time series
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
-    >>> lcss_pairwise_distance(X)
-    array([[0.        , 0.66666667, 1.        ],
-           [0.66666667, 0.        , 0.66666667],
-           [1.        , 0.66666667, 0.        ]])
+    >>> edr_pairwise_distance(X)
+    array([[0., 1., 1.],
+           [1., 0., 1.],
+           [1., 1., 0.]])
 
     >>> # Distance between two collections of time series
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
     >>> y = np.array([[[11, 12, 13]],[[14, 15, 16]], [[17, 18, 19]]])
-    >>> lcss_pairwise_distance(X, y)
+    >>> edr_pairwise_distance(X, y)
     array([[1., 1., 1.],
            [1., 1., 1.],
            [1., 1., 1.]])
 
     >>> X = np.array([[[1, 2, 3]],[[4, 5, 6]], [[7, 8, 9]]])
     >>> y_univariate = np.array([11, 12, 13])
-    >>> lcss_pairwise_distance(X, y_univariate)
+    >>> edr_pairwise_distance(X, y_univariate)
     array([[1.],
            [1.],
            [1.]])
 
     >>> # Distance between each TS in a collection of unequal-length time series
     >>> X = [np.array([1, 2, 3]), np.array([4, 5, 6, 7]), np.array([8, 9, 10, 11, 12])]
-    >>> lcss_pairwise_distance(X)
-    array([[0.        , 0.66666667, 1.        ],
-           [0.66666667, 0.        , 0.75      ],
-           [1.        , 0.75      , 0.        ]])
+    >>> edr_pairwise_distance(X)
+    array([[0.  , 0.75, 0.6 ],
+           [0.75, 0.  , 0.8 ],
+           [0.6 , 0.8 , 0.  ]])
     """
     multivariate_conversion = _is_multivariate(X, y)
     _X, unequal_length = _convert_to_list(X, "X", multivariate_conversion)
+
     if y is None:
         # To self
-        return _lcss_pairwise_distance(
+        return _edr_pairwise_distance(
             _X, window, epsilon, itakura_max_slope, unequal_length
         )
+
     _y, unequal_length = _convert_to_list(y, "y", multivariate_conversion)
-    return _lcss_from_multiple_to_multiple_distance(
+    return _edr_from_multiple_to_multiple_distance(
         _X, _y, window, epsilon, itakura_max_slope, unequal_length
     )
 
 
 @njit(cache=True, fastmath=True)
-def _lcss_pairwise_distance(
+def _edr_pairwise_distance(
     X: NumbaList[np.ndarray],
     window: Optional[float],
-    epsilon: float,
+    epsilon: Optional[float],
     itakura_max_slope: Optional[float],
     unequal_length: bool,
 ) -> np.ndarray:
     n_cases = len(X)
     distances = np.zeros((n_cases, n_cases))
+
     if not unequal_length:
         n_timepoints = X[0].shape[1]
         bounding_matrix = create_bounding_matrix(
@@ -329,18 +340,18 @@ def _lcss_pairwise_distance(
                 bounding_matrix = create_bounding_matrix(
                     x1.shape[1], x2.shape[1], window, itakura_max_slope
                 )
-            distances[i, j] = _lcss_distance(x1, x2, bounding_matrix, epsilon)
+            distances[i, j] = _edr_distance(x1, x2, bounding_matrix, epsilon)
             distances[j, i] = distances[i, j]
 
     return distances
 
 
 @njit(cache=True, fastmath=True)
-def _lcss_from_multiple_to_multiple_distance(
+def _edr_from_multiple_to_multiple_distance(
     x: NumbaList[np.ndarray],
     y: NumbaList[np.ndarray],
     window: Optional[float],
-    epsilon: float,
+    epsilon: Optional[float],
     itakura_max_slope: Optional[float],
     unequal_length: bool,
 ) -> np.ndarray:
@@ -359,19 +370,19 @@ def _lcss_from_multiple_to_multiple_distance(
                 bounding_matrix = create_bounding_matrix(
                     x1.shape[1], y1.shape[1], window, itakura_max_slope
                 )
-            distances[i, j] = _lcss_distance(x1, y1, bounding_matrix, epsilon)
+            distances[i, j] = _edr_distance(x1, y1, bounding_matrix, epsilon)
     return distances
 
 
 @njit(cache=True, fastmath=True)
-def lcss_alignment_path(
+def edr_alignment_path(
     x: np.ndarray,
     y: np.ndarray,
     window: Optional[float] = None,
-    epsilon: float = 1.0,
+    epsilon: Optional[float] = None,
     itakura_max_slope: Optional[float] = None,
 ) -> tuple[list[tuple[int, int]], float]:
-    """Compute the LCSS alignment path between two time series.
+    """Compute the EDR alignment path between two time series.
 
     Parameters
     ----------
@@ -382,9 +393,10 @@ def lcss_alignment_path(
     window : float, default=None
         The window to use for the bounding matrix. If None, no bounding matrix
         is used.
-    epsilon : float, default=1.
+    epsilon : float, default=None
         Matching threshold to determine if two subsequences are considered close
-        enough to be considered 'common'. The default is 1.
+        enough to be considered 'common'. If not specified as per the original paper
+        epsilon is set to a quarter of the maximum standard deviation.
     itakura_max_slope : float, default=None
         Maximum slope as a proportion of the number of time points used to create
         Itakura parallelogram on the bounding matrix. Must be between 0. and 1.
@@ -396,7 +408,7 @@ def lcss_alignment_path(
         of the index in x and the index in y that have the best alignment according
         to the cost matrix.
     float
-        The LCSS distance between the two time series.
+        The EDR distance between the two time series.
 
     Raises
     ------
@@ -406,30 +418,14 @@ def lcss_alignment_path(
     Examples
     --------
     >>> import numpy as np
-    >>> from aeon.distances import lcss_alignment_path
+    >>> from aeon.distances import edr_alignment_path
     >>> x = np.array([[1, 2, 3, 6]])
     >>> y = np.array([[1, 2, 3, 4]])
-    >>> path, dist = lcss_alignment_path(x, y)
-    >>> path
-    [(0, 0), (1, 1), (2, 2)]
+    >>> edr_alignment_path(x, y)
+    ([(0, 0), (1, 1), (2, 2), (3, 3)], 0.25)
     """
-    x_size = x.shape[-1]
-    y_size = y.shape[-1]
-    bounding_matrix = create_bounding_matrix(x_size, y_size, window, itakura_max_slope)
-    if x.ndim == 1 and y.ndim == 1:
-        _x = x.reshape((1, x.shape[0]))
-        _y = y.reshape((1, y.shape[0]))
-        cost_matrix = _lcss_cost_matrix(_x, _y, bounding_matrix, epsilon)
-        distance = 1 - (float(cost_matrix[x_size, y_size] / min(x_size, y_size)))
-        return (
-            compute_lcss_return_path(_x, _y, epsilon, bounding_matrix, cost_matrix),
-            distance,
-        )
-    if x.ndim == 2 and y.ndim == 2:
-        cost_matrix = _lcss_cost_matrix(x, y, bounding_matrix, epsilon)
-        distance = 1 - (float(cost_matrix[x_size, y_size] / min(x_size, y_size)))
-        return (
-            compute_lcss_return_path(x, y, epsilon, bounding_matrix, cost_matrix),
-            distance,
-        )
-    raise ValueError("x and y must be 1D or 2D arrays")
+    cost_matrix = edr_cost_matrix(x, y, window, epsilon, itakura_max_slope)
+    return (
+        compute_min_return_path(cost_matrix),
+        cost_matrix[x.shape[-1] - 1, y.shape[-1] - 1] / max(x.shape[-1], y.shape[-1]),
+    )
