+++ aeon/distances/elastic/_twe.py
@@ -8,10 +8,10 @@ import numpy as np
 from numba import njit
 from numba.typed import List as NumbaList
 
-from aeon.distances._alignment_paths import compute_min_return_path
-from aeon.distances._bounding_matrix import create_bounding_matrix
 from aeon.distances._euclidean import _univariate_euclidean_distance
 from aeon.distances._utils import _convert_to_list, _is_multivariate
+from aeon.distances.elastic._alignment_paths import compute_min_return_path
+from aeon.distances.elastic._bounding_matrix import create_bounding_matrix
 
 
 @njit(cache=True, fastmath=True)
