+++ aeon/distances/elastic/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Elastic distance tests."""
